requirements.txt
google-cloud-aiplatform==1.69.0
python-dotenv==1.0.1